fix: attempt to make table base component
